pyproject.toml
[project]
name = "mobile-tts-yds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=3.6.0,<4.0.0",
    "faster-whisper>=1.2.1",
    "hydra-core>=1.3.2",
    "jiwer>=4.0.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "omegaconf>=2.3.0",
    "onnx>=1.21.0",
    "onnxconverter-common>=1.16.0",
    "onnxruntime>=1.24.4",
    "onnxruntime-extensions>=0.15.2",
    "onnxruntime-gpu>=1.24.4",
    "onnxruntime-tools>=1.7.0",
    "onnxscript>=0.7.0",
    "optimum[onnxruntime]>=2.1.0",
    "qwen-tts>=0.1.1",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "torchcodec>=0.11.0",
    "torchmetrics>=1.8.2",
]
